src: rename (issue #21)
